:sparkles: restructure project layout and dynamic automated_actions_cli
