--- automated_actions/api/v1/noop/models.py
@@ -1,9 +1,0 @@
-from pydantic import BaseModel
-
-
-class NoopParam(BaseModel):
-    alias: str
-    text: str
-
-    # pydantic config
-    model_config = {"extra": "ignore"}
