--- automated_actions_cli/automated_actions_cli/__main__.py
@@ -1,9 +1,0 @@
-import logging
-
-from .cli import app
-
-log = logging.getLogger("automated_actions_cli")
-
-
-if __name__ == "__main__":
-    app()
