--- automated_actions_client/automated_actions_client/api/v1/run_foobar_action.py
@@ -1,161 +1,0 @@
-from http import HTTPStatus
-from typing import Any
-
-import httpx
-
-from ... import errors
-from ...client import AuthenticatedClient, Client
-from ...models.http_validation_error import HTTPValidationError
-from ...types import Response
-
-
-def _get_kwargs(
-    pk: str,
-) -> dict[str, Any]:
-    _kwargs: dict[str, Any] = {
-        "method": "get",
-        "url": f"/api/v1/foobar/{pk}",
-    }
-
-    return _kwargs
-
-
-def _parse_response(
-    *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> Any | HTTPValidationError | None:
-    if response.status_code == 200:
-        response_200 = response.json()
-        return response_200
-    if response.status_code == 422:
-        response_422 = HTTPValidationError.from_dict(response.json())
-
-        return response_422
-    if client.raise_on_unexpected_status:
-        raise errors.UnexpectedStatus(response.status_code, response.content)
-    else:
-        return None
-
-
-def _build_response(
-    *, client: AuthenticatedClient | Client, response: httpx.Response
-) -> Response[Any | HTTPValidationError]:
-    return Response(
-        status_code=HTTPStatus(response.status_code),
-        content=response.content,
-        headers=response.headers,
-        parsed=_parse_response(client=client, response=response),
-    )
-
-
-def sync_detailed(
-    pk: str,
-    *,
-    client: AuthenticatedClient | Client,
-) -> Response[Any | HTTPValidationError]:
-    """Run a foobar action
-
-     Noop action
-
-    Args:
-        pk (str):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Union[Any, HTTPValidationError]]
-    """
-
-    kwargs = _get_kwargs(
-        pk=pk,
-    )
-
-    response = client.get_httpx_client().request(
-        **kwargs,
-    )
-
-    return _build_response(client=client, response=response)
-
-
-def sync(
-    pk: str,
-    *,
-    client: AuthenticatedClient | Client,
-) -> Any | HTTPValidationError | None:
-    """Run a foobar action
-
-     Noop action
-
-    Args:
-        pk (str):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Union[Any, HTTPValidationError]
-    """
-
-    return sync_detailed(
-        pk=pk,
-        client=client,
-    ).parsed
-
-
-async def asyncio_detailed(
-    pk: str,
-    *,
-    client: AuthenticatedClient | Client,
-) -> Response[Any | HTTPValidationError]:
-    """Run a foobar action
-
-     Noop action
-
-    Args:
-        pk (str):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Response[Union[Any, HTTPValidationError]]
-    """
-
-    kwargs = _get_kwargs(
-        pk=pk,
-    )
-
-    response = await client.get_async_httpx_client().request(**kwargs)
-
-    return _build_response(client=client, response=response)
-
-
-async def asyncio(
-    pk: str,
-    *,
-    client: AuthenticatedClient | Client,
-) -> Any | HTTPValidationError | None:
-    """Run a foobar action
-
-     Noop action
-
-    Args:
-        pk (str):
-
-    Raises:
-        errors.UnexpectedStatus: If the server returns an undocumented status code and Client.raise_on_unexpected_status is True.
-        httpx.TimeoutException: If the request takes longer than Client.timeout.
-
-    Returns:
-        Union[Any, HTTPValidationError]
-    """
-
-    return (
-        await asyncio_detailed(
-            pk=pk,
-            client=client,
-        )
-    ).parsed
