--- automated_actions_client/automated_actions_client/models/noop_param.py
@@ -1,64 +1,0 @@
-from typing import Any, TypeVar
-
-from attrs import define as _attrs_define
-from attrs import field as _attrs_field
-
-T = TypeVar("T", bound="NoopParam")
-
-
-@_attrs_define
-class NoopParam:
-    """
-    Attributes:
-        alias (str):
-        text (str):
-    """
-
-    alias: str
-    text: str
-    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
-
-    def to_dict(self) -> dict[str, Any]:
-        alias = self.alias
-
-        text = self.text
-
-        field_dict: dict[str, Any] = {}
-        field_dict.update(self.additional_properties)
-        field_dict.update({
-            "alias": alias,
-            "text": text,
-        })
-
-        return field_dict
-
-    @classmethod
-    def from_dict(cls: type[T], src_dict: dict[str, Any]) -> T:
-        d = src_dict.copy()
-        alias = d.pop("alias")
-
-        text = d.pop("text")
-
-        noop_param = cls(
-            alias=alias,
-            text=text,
-        )
-
-        noop_param.additional_properties = d
-        return noop_param
-
-    @property
-    def additional_keys(self) -> list[str]:
-        return list(self.additional_properties.keys())
-
-    def __getitem__(self, key: str) -> Any:
-        return self.additional_properties[key]
-
-    def __setitem__(self, key: str, value: Any) -> None:
-        self.additional_properties[key] = value
-
-    def __delitem__(self, key: str) -> None:
-        del self.additional_properties[key]
-
-    def __contains__(self, key: str) -> bool:
-        return key in self.additional_properties
