+++ packages/automated_actions/README.md
@@ -1,0 +1,47 @@
+# Automated Actions
+
+[![Ruff](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/ruff/main/assets/badge/v2.json)](https://github.com/astral-sh/ruff)
+[![uv](https://img.shields.io/endpoint?url=https://raw.githubusercontent.com/astral-sh/uv/main/assets/badge/v0.json)](https://github.com/astral-sh/uv)
+
+Welcome to Automated Actions! :rocket:
+
+## Features
+
+TODO
+
+## Deployment
+
+See `openshift` templates.
+
+## Usage
+
+```bash
+automated-actions
+```
+
+## Development
+
+Create and maintain a development virtual environment via:
+
+```bash
+make dev-env
+```
+
+Activate the virtual environment in the current shell via:
+
+```bash
+source .venv/bin/activate
+```
+
+Run the tests suite via:
+
+```bash
+make tests
+```
+
+[Api docs](http://localhost:8080/redoc)
+[Swagger](http://localhost:8080/doc)
+
+## License
+
+This project is licensed under the terms of the [Apache 2.0 license](/LICENSE).
