+++ packages/automated_actions/automated_actions/api/v1/noop/models.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class NoopParam(BaseModel):
+    alias: str
+    description: str = "no description"
+
+    # pydantic config
+    model_config = {"extra": "ignore"}
