+++ packages/automated_actions/tests/test_stub.py
@@ -1,0 +1,2 @@
+def test_automated_actions_stub() -> None:
+    assert True
