+++ packages/automated_actions_cli/automated_actions_cli/__main__.py
@@ -1,0 +1,9 @@
+import logging
+
+from automated_actions_cli.cli import app
+
+log = logging.getLogger("automated_actions_cli")
+
+
+if __name__ == "__main__":
+    app()
