+++ packages/automated_actions_cli/tests/test_stub.py
@@ -1,0 +1,2 @@
+def test_automated_actions_cli_stub() -> None:
+    assert True
