+++ packages/automated_actions_client/automated_actions_client/api/v1/__init__.py
@@ -1,0 +1,6 @@
+from . import foobar, noop
+
+__all__ = [
+    "foobar",
+    "noop",
+]
