+++ packages/automated_actions_client/automated_actions_client/models/noop_param.py
@@ -1,0 +1,67 @@
+from typing import Any, TypeVar
+
+from attrs import define as _attrs_define
+from attrs import field as _attrs_field
+
+from ..types import UNSET, Unset
+
+T = TypeVar("T", bound="NoopParam")
+
+
+@_attrs_define
+class NoopParam:
+    """
+    Attributes:
+        alias (str):
+        description (Union[Unset, str]):  Default: 'no description'.
+    """
+
+    alias: str
+    description: Unset | str = "no description"
+    additional_properties: dict[str, Any] = _attrs_field(init=False, factory=dict)
+
+    def to_dict(self) -> dict[str, Any]:
+        alias = self.alias
+
+        description = self.description
+
+        field_dict: dict[str, Any] = {}
+        field_dict.update(self.additional_properties)
+        field_dict.update({
+            "alias": alias,
+        })
+        if description is not UNSET:
+            field_dict["description"] = description
+
+        return field_dict
+
+    @classmethod
+    def from_dict(cls: type[T], src_dict: dict[str, Any]) -> T:
+        d = src_dict.copy()
+        alias = d.pop("alias")
+
+        description = d.pop("description", UNSET)
+
+        noop_param = cls(
+            alias=alias,
+            description=description,
+        )
+
+        noop_param.additional_properties = d
+        return noop_param
+
+    @property
+    def additional_keys(self) -> list[str]:
+        return list(self.additional_properties.keys())
+
+    def __getitem__(self, key: str) -> Any:
+        return self.additional_properties[key]
+
+    def __setitem__(self, key: str, value: Any) -> None:
+        self.additional_properties[key] = value
+
+    def __delitem__(self, key: str) -> None:
+        del self.additional_properties[key]
+
+    def __contains__(self, key: str) -> bool:
+        return key in self.additional_properties
