+++ packages/automated_actions_client/pyproject.toml
@@ -1,0 +1,60 @@
+[project]
+name = "automated-actions-client"
+version = "0.1.0"
+description = "Automated Actions Client"
+authors = [
+    # Feel free to add or change authors
+    { name = "AppSRE", email = "sd-app-sre@redhat.com" },
+]
+license = { text = "Apache 2.0" }
+readme = "README.md"
+requires-python = "~= 3.12.0"
+dependencies = ["httpx", "attrs"]
+
+[project.urls]
+homepage = "https://github.com/app-sre/automated-actions"
+repository = "https://github.com/app-sre/automated-actions"
+documentation = "https://github.com/app-sre/automated-actions"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.hatch.build.targets.sdist]
+only-include = ["automated_actions_client"]
+
+[tool.hatch.build.targets.wheel]
+only-include = ["automated_actions_client"]
+
+# Ruff configuration
+[tool.ruff]
+line-length = 88
+src = ["automated_actions_client"]
+fix = true
+
+[tool.ruff.lint]
+select = ["F", "I", "UP", "RUF022"]
+
+[tool.ruff.format]
+preview = true
+
+[tool.ruff.lint.isort]
+known-first-party = ["automated_actions_client"]
+
+[tool.mypy]
+files = ["automated_actions_client"]
+enable_error_code = ["truthy-bool", "redundant-expr"]
+no_implicit_optional = true
+check_untyped_defs = true
+warn_unused_ignores = true
+show_error_codes = true
+disallow_untyped_defs = true
+disallow_incomplete_defs = true
+
+# Coverage configuration
+[tool.coverage.run]
+branch = true
+omit = ["*/tests/*"]
+
+# [tool.coverage.report]
+# fail_under = 90
