+++ packages/automated_actions_client/tests/test_stub.py
@@ -1,0 +1,2 @@
+def test_automated_actions_client_stub() -> None:
+    assert True
