+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "automated-actions"
+name = "automated-actions-project"
 version = "0.1.0"
 description = "Automated Actions Server"
 authors = [
@@ -9,26 +9,16 @@ authors = [
 license = { text = "Apache 2.0" }
 readme = "README.md"
 requires-python = "~= 3.12.0"
-dependencies = [
-    "celery[sqs]>=5.4.0",
-    "fastapi>=0.115.8",
-    "prometheus-fastapi-instrumentator>=7.0.2",
-    "pydantic>=2.10.6",
-    "pydantic-settings>=2.7.1",
-    "uvicorn>=0.34.0",
-]
 
 [tool.uv.sources]
+automated-actions = { workspace = true }
 automated-actions-client = { workspace = true }
 automated-actions-cli = { workspace = true }
+# fix for https://github.com/openapi-generators/openapi-python-client/issues/1188
+openapi-python-client = { git = "https://github.com/chassing/openapi-python-client", branch = "issue1188-models-enums-iterable" }
 
 [tool.uv.workspace]
-members = ["automated_actions_client", "automated_actions_cli"]
-
-[project.urls]
-homepage = "https://github.com/app-sre/automated-actions"
-repository = "https://github.com/app-sre/automated-actions"
-documentation = "https://github.com/app-sre/automated-actions"
+members = ["packages/*"]
 
 [dependency-groups]
 dev = [
@@ -38,101 +28,4 @@ dev = [
     "pytest ~=8.2.0",
     "pytest-cov ~=6.0.0",
     "openapi-python-client>=0.23.1",
-    "automated-actions-client",
-    "automated-actions-cli",
 ]
-
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[tool.hatch.build.targets.sdist]
-only-include = ["automated_actions"]
-
-[tool.hatch.build.targets.wheel]
-only-include = ["automated_actions"]
-
-# Ruff configuration
-[tool.ruff]
-line-length = 88
-src = ["automated_actions"]
-extend-exclude = [
-    # exclude some common cache and tmp directories
-    ".local",
-    ".cache",
-    "tmp",
-]
-fix = true
-
-[tool.ruff.lint]
-preview = true
-select = ["ALL"]
-ignore = [
-    "CPY",     # Missing copyright header
-    "D100",    # Missing docstring in public module
-    "D101",    # Missing docstring in public class
-    "D102",    # Missing docstring in public method
-    "D103",    # Missing docstring in public function
-    "D104",    # Missing docstring in public package
-    "D105",    # Missing docstring in magic method
-    "D107",    # Missing docstring in __init__
-    "D203",    # 1 blank line required before class docstring
-    "D211",    # No blank lines allowed before class docstring
-    "D212",    # multi-line-summary-first-line
-    "D213",    # multi-line-summary-second-line
-    "D4",      # Doc string style
-    "E501",    # Line too long
-    "G004",    # Logging statement uses f-string
-    "PLR0904", # Too many public methods
-    "PLR0913", # Too many arguments
-    "PLR0917", # Too many positional arguments
-    "S101",    # Use of assert detected. Pytest uses assert
-    "S404",    # subprocess import
-    "EM101",   # Exception must not use a string literal, assign to variable first
-    "EM102",   # Exception must not use an f-string literal, assign to variable first
-    "TRY003",  # Avoid specifying long messages outside the exception class
-    "TRY300",  # try-consider-else
-    # pydoclint
-    "DOC",
-    # May conflict with the formatter, https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules
-    "W191",
-    "E111",
-    "E114",
-    "E117",
-    "D206",
-    "D300",
-    "Q",
-    "COM812",
-    "COM819",
-    "ISC001",
-]
-[tool.ruff.format]
-preview = true
-
-[tool.ruff.lint.isort]
-known-first-party = ["automated_actions"]
-
-# Mypy configuration
-[tool.mypy]
-files = ["automated_actions", "tests"]
-enable_error_code = ["truthy-bool", "redundant-expr"]
-no_implicit_optional = true
-check_untyped_defs = true
-warn_unused_ignores = true
-show_error_codes = true
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
-
-[[tool.mypy.overrides]]
-# Below are all of the packages that don't implement stub packages. Mypy will throw an error if we don't ignore the
-# missing imports. See: https://mypy.readthedocs.io/en/stable/running_mypy.html#missing-imports
-module = ["celery.*"]
-ignore_missing_imports = true
-
-# Coverage configuration
-[tool.coverage.run]
-branch = true
-omit = ["*/tests/*"]
-
-[tool.coverage.report]
-fail_under = 90
